Moved routes file and moved eligiblity API to routes
