+++ next-frontend/src/app/map/page.jsx
@@ -12,7 +12,7 @@ import {
   SelectValue,
 } from "@/components/ui/select"
 import { Input } from "@/components/ui/input"
-import { getMapData } from "@/app/api/mapData/route";
+import { getMapData } from "@/app/api/route";
 import { Popover, PopoverTrigger, PopoverContent } from "@/components/ui/popover";
 import { Checkbox } from "@/components/ui/checkbox";
 import debounce from "lodash.debounce";
